feat(3d): baked-tile terrain engine — detail layer, PL/SK DEM merge, overlay seating on baked tiles
Terrain quality (the "smooth runways/ratrak" bug, docs/SMOOTH-SURFACE-BUG.md):
- BDT2 tile format: per-cell residual RMS the box-average downsample discards,
  carried as a per-vertex detail channel (backward compatible with BDT1)
- fragment shader bends the shading normal by the REAL residual amplitude;
  triplanar noise-plane pick (no fall-line grooves) and hashT lattice wrap
  (sin() precision past ±8192 rad read as banding)
- ring/live paths compute the residual on the fly (BuildAdaptiveTiles had no
  detailGrid); strided window keeps it O(1) per vertex; distance fade so far
  tiles never out-bump near ones; NativeMicroDetail adds a capped micro-bump
  on native z16 where 1 m LiDAR cannot resolve sub-metre texture

PL/SK border ("rock cut in half" seam near Mięguszowieckie):
- per-pixel merge of SK DMR5 into partially-void gugik z16 tiles (zero-filled
  and NaN/-999 halves), 1746 tiles / ~105M px; quadtree all-children-baked no
  longer collapses border quads to coarse z15

Overlays glued to the rendered surface:
- trails/route/roads/user dot seat on the REAL baked z16 elevation
  (BakedElevationIndex) instead of the static coarse base
- trail decal: camera-target-centred mask window (was whole-mesh under baked
  streaming), pixel-floor line width; trail lines alpha-blend so decal shows
  through lost z-fights; route-film dot/camera follow the conflated track
- ortho: per-cell near/far resolution tier by camera distance with hysteresis
  (OrthoDistanceTier) instead of a flat 4096 px cap

P1/P2 groundwork: 32-bit mesh indices (single-mesh 256² baked tiles), residency
budget 512MB→1280MB, GPU pass timers, base-occlusion planner (gated off).

+++ src/MapaTur.Application/Terrain/Terrain3DProjection.cs
@@ -138,7 +138,10 @@ public static class Terrain3DProjection
             }
         }
 
-        ushort[] sourceIndices = mesh.Indices;
+        // The mesh's index buffer is 32-bit (uint). This Skia fallback emits ushort SKVertices indices, which can
+        // only address ≤ 65 536 verts — fine here because the live tiles this path renders stay small (≤ 250²).
+        // The baked-streaming tiles (which can exceed that) go through the GL renderer, not this CPU fallback.
+        uint[] sourceIndices = mesh.Indices;
         int triangleCount = sourceIndices.Length / 3;
         scratch.EnsureTriangleCapacity(triangleCount);
         float[] depths = scratch.Depths;
@@ -150,9 +153,9 @@ public static class Terrain3DProjection
         for (int t = 0; t < triangleCount; t++)
         {
             int baseIdx = t * 3;
-            ushort i0 = sourceIndices[baseIdx];
-            ushort i1 = sourceIndices[baseIdx + 1];
-            ushort i2 = sourceIndices[baseIdx + 2];
+            int i0 = (int)sourceIndices[baseIdx];
+            int i1 = (int)sourceIndices[baseIdx + 1];
+            int i2 = (int)sourceIndices[baseIdx + 2];
 
             Vector3 v0 = screen[i0];
             Vector3 v1 = screen[i1];
@@ -222,9 +225,9 @@ public static class Terrain3DProjection
             int t = triangleIndices[k];
             int baseSrc = t * 3;
             int baseDst = slot * 3;
-            visibleIndices[baseDst] = sourceIndices[baseSrc];
-            visibleIndices[baseDst + 1] = sourceIndices[baseSrc + 1];
-            visibleIndices[baseDst + 2] = sourceIndices[baseSrc + 2];
+            visibleIndices[baseDst] = (ushort)sourceIndices[baseSrc];
+            visibleIndices[baseDst + 1] = (ushort)sourceIndices[baseSrc + 1];
+            visibleIndices[baseDst + 2] = (ushort)sourceIndices[baseSrc + 2];
         }
 
         return new ProjectedTerrainFrame(screen, visibleIndices, vertexCount, visibleCount * 3);
